Fix for #11
If JoinReportHtmlOptions.OtherColumnsToSelect are used, rows are generated with every PivotProperty and StatisticColumn combination possible even if the values have no correlation in the initial benchmark data.
OtherColumnsToSelect should act as additional group key but without splitting the table. The design of the result table leads to this conclusion.
With this fix correct pivot tables are shown.